评论功能改成弹窗，可上传图片

+++ src/main/java/com/wei/dao/UsertagDao.java
@@ -23,6 +23,6 @@ public interface UsertagDao extends BaseMapper<Usertag> {
      * @return 标签集合
      */
     @Select("select u.uTid,u.uid,t.tagId,t.tagName,count(a.tagId) count from userTag u join tag t on t.tagId=u.tagId join article a on a.tagId=u.tagId where u.uid=#{uid} group by u.tagId")
-    List<UsertagVo> selectTagsByUid(Integer uid);
+    List<UsertagVo> selectTagsByUid(long uid);
 
 }
